Built out items and tasks
Fixed job route, updated models to reflect an easier way of keeping track of jobs : user. Before we were adding an array of jobs to the user model, now we are adding a user by grabbing the id from the auth token and passing it into every new job so the job will be created based on who's logged in. This technique was found via stackoverflow and I thought it was a better way to set up the models.